requirements.txt
wheel
requests
ffmpeg-python
mutagen
music-tag